refactor: separate controller and application logic
