--- cmd/currency-converter/router/router.go
@@ -1,17 +1,0 @@
-package router
-
-import (
-	"github.com/gin-gonic/gin"
-	"log"
-	"net/http"
-)
-
-func SetupRoutes(router *gin.Engine) {
-	router.GET("/ping", ping)
-}
-
-func ping(context *gin.Context) {
-	log.Println("PONG!")
-
-	context.String(http.StatusOK, "pong")
-}
